Add test fixtures with zip jsonl
