+++ tests/test_load.py
@@ -157,17 +157,6 @@ def metric_loading_script_dir(tmp_path):
     return str(script_dir)
 
 
-@pytest.fixture(scope="session")
-def zip_csv_with_dir_path(csv_path, csv2_path, tmp_path_factory):
-    import zipfile
-
-    path = tmp_path_factory.mktemp("data") / "dataset_with_dir.csv.zip"
-    with zipfile.ZipFile(path, "w") as f:
-        f.write(csv_path, arcname=os.path.join("main_dir", os.path.basename(csv_path)))
-        f.write(csv2_path, arcname=os.path.join("main_dir", os.path.basename(csv2_path)))
-    return path
-
-
 @pytest.mark.parametrize("data_file, expected_module", [("zip_csv_path", "csv"), ("zip_csv_with_dir_path", "csv")])
 def test_infer_module_for_data_files_in_archives(data_file, expected_module, zip_csv_path, zip_csv_with_dir_path):
     data_file_paths = {"zip_csv_path": zip_csv_path, "zip_csv_with_dir_path": zip_csv_with_dir_path}
